pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "se4x-companion"
version = "0.1.0"
description = "Companion app for Space Empires 4X board game"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1",
    "flask-wtf>=1.2",
    "flask-babel>=4.0",
    "python-dotenv>=1.1",
    "flask-sqlalchemy>=3.1.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "prek>=0.3",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
filterwarnings = ["ignore::ResourceWarning"]

[project.scripts]
se4x = "app:main"
